// src/types/types.ts
export interface Category {
  id: number;
  title: string;
  name?: string; // Name if 'title' is unavailable
  slug: string;
  image: string;
}

export interface FetchCategoriesPayload {
  results: Category[];
}

export interface CategoryState {
  categories: Category[];
  status: "idle" | "loading" | "succeeded" | "failed";
  error: string | null;
}

export interface User {
  id: number;
  username: string;
  email: string;
}
export interface Store {
  id: number;
  name: string;
  rating: number;
  reviews_count: number;
  location: string;
  overview: string;
  image: string;
  phone_number?: string;
  whats_app?: string;
  working_hours: string | null;
  is_verified: boolean;
  is_responsive: boolean;
  event_planning_categories: string[];
  rent_hire_categories: string[];
}